backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
asyncpg==0.30.0
sqlalchemy==2.0.35
python-jose==3.3.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
bcrypt==4.2.0
slowapi==0.1.9
redis==5.0.8
email-validator==2.1.1
